Improved hit detection + Can land attack now
You can land attack if a player walks into your attack animation while
you are performing it before this was not the case. Now it's a little
more realistic.

Tried to put in Jo's code again no luck as of yet.
Fixed sticky issue with special attack (But now if you attack a player
in special attack they are forced to release their special and hurt you
- makes you think before spamming) Though the flinch animation is not
working correctly for that case