Added 2-Player map voting
Determines which map had the most votes. If there's a tie, it randomly
chooses between the highest voted maps. If the random map was chosen,
then it randomly chooses.

For example, if P1 votes Lava and P2 votes Random, it will choose 50/50
between Lava and Random. Then if Random wins, it will randomly choose
one of the four maps. So there's actually a 5/8 chance of it being Lava.